pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[tool.maturin]
module-name = "_formatparse"
python-source = "."
bindings = "pyo3"
compatibility = "linux"

[project]
name = "formatparse"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
description = "Parse strings using a specification based on the Python format() syntax (Rust implementation)"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Odos Matthews", email = "odosmatthews@gmail.com" }]
keywords = ["parse", "format", "string", "parsing"]
dependencies = []

[project.optional-dependencies]
test = [
    "hypothesis>=6.0",
    "pytest-cov>=4.0",
    "pytest-benchmark>=4.0",
    "pympler>=0.9",
    "memory_profiler>=0.60",
]
dev = ["mutmut>=2.0"]

[project.urls]
Homepage = "https://github.com/eddiethedean/formatparse"
Repository = "https://github.com/eddiethedean/formatparse"

[tool.pytest.ini_options]
markers = [
    "benchmark: marks tests as benchmarks",
    "slow: marks tests as slow running",
    "stress: marks tests as stress/load tests",
]

[tool.coverage.run]
source = ["formatparse"]
omit = ["tests/*", "*/tests/*"]

[tool.coverage.report]
fail_under = 90
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

